bot1/requirements.txt
python-telegram-bot==20.7
openai==1.3.5
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
logging==0.4.9.6 